# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pets App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/pins": {
            "get": {
                "tags": [
                    "Pins"
                ],
                "summary": "Get list of the user`s pins.",
                "operationId": "62dc87663d3ee1cf3519966cff94c098",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResourceCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pins"
                ],
                "summary": "Create a new pin.",
                "operationId": "bbfae9748d7dfe23ea2600d341dcd679",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PinStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PinFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pins/search": {
            "get": {
                "tags": [
                    "Pins"
                ],
                "summary": "Get list of pins matched by filters.",
                "operationId": "cbc1db2dd6485777945d4431c14e066b",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "path",
                        "required": false
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "required": false
                    },
                    {
                        "name": "radius",
                        "in": "path",
                        "required": false
                    },
                    {
                        "name": "type_ids[]",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PinShortResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pins/{pin}": {
            "get": {
                "tags": [
                    "Pins"
                ],
                "summary": "Get a pin.",
                "operationId": "a0b09c30b0994b5e679f4de20922c82c",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PinFullResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pins"
                ],
                "summary": "Delete a pin.",
                "operationId": "7c6960ae78c7d9549007e9aadaad5ac2",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Pins"
                ],
                "summary": "Update a pin.",
                "operationId": "bbac96604f5c26375e38474f3c73497c",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PinStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PinFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user details.",
                "operationId": "840711b2cfe9f9c00a12c154bfb21787",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserFullResource"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile.",
                "operationId": "2416ea03dc2ac6a315f36e2de0dd8f71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "expample@mail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "380009998877"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserFullResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/profile/avatar": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user avatar.",
                "operationId": "aab356471782aa4ab172229bb1e61fc6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "avatar": {
                                                "description": "User avatar",
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "thumb": {
                                            "type": "string"
                                        },
                                        "full": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PinStoreRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Some title"
                    },
                    "type_id": {
                        "type": "string",
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "contact": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginatedResourceCollection": {
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/PinShortResource"
                                }
                            ]
                        }
                    },
                    "meta": {
                        "properties": {
                            "total": {
                                "type": "integer"
                            },
                            "current": {
                                "type": "integer"
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PinFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "contact": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    },
                    "gallery": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "PinShortResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "UserFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "thumb": {
                                    "type": "string"
                                },
                                "full": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}